Replace Auth0 code with custom authentication
