+++ app/src/auth/index.ts
@@ -1,2 +1,2 @@
 export {Auth0ProviderWithHistory} from './Auth0ProviderWithHistory';
-export {ProtectedRoute} from './ProtectedRoute';
+export {ProtectedRoute} from './Auth0ProtectedRoute';
