+++ app/src/views/index.ts
@@ -5,4 +5,3 @@ export {Home} from './Home';
 export {SignIn} from './SignIn';
 export {SignUp} from './SignUp';
 export {AccountVerification} from './AccountVerification';
-export {PrivateView} from './PrivateView';
